backend/requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.3; python_version >= '3.8'
aiohttp==3.11.7; python_version >= '3.9'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.6.2.post1; python_version >= '3.9'
attrs==24.2.0; python_version >= '3.7'
azure-ai-documentintelligence==1.0.0b3; python_version >= '3.8'
azure-common==1.1.28
azure-core==1.32.0; python_version >= '3.8'
azure-cosmos==4.9.0; python_version >= '3.8'
azure-identity==1.19.0; python_version >= '3.8'
azure-search-documents==11.5.2; python_version >= '3.8'
azure-storage-blob==12.22.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==43.0.3; python_version >= '3.7'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
faiss-cpu==1.9.0.post1; python_version >= '3.9'
flask==3.1.0; python_version >= '3.9'
flask-cors==5.0.0
flask-sqlalchemy==3.1.1; python_version >= '3.8'
flatbuffers==24.3.25
frozenlist==1.5.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.10; python_version >= '3.6'
isodate==0.7.2; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
jiter==0.7.1; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
langchain==0.2.17; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.2.6; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.2.43; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-ollama==0.1.3; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-openai==0.1.25; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-text-splitters==0.2.4; python_version < '4.0' and python_full_version >= '3.8.1'
langchainhub==0.1.21; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.1.144; python_version < '4.0' and python_full_version >= '3.8.1'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.23.1; python_version >= '3.9'
mpmath==1.3.0
msal==1.31.1; python_version >= '3.7'
msal-extensions==1.2.0; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.4; python_version >= '3.12'
ollama==0.4.0; python_version >= '3.8' and python_version < '4.0'
onnxruntime==1.20.1
openai==1.55.0; python_version >= '3.8'
opencv-python==4.10.0.84; python_version >= '3.6'
orjson==3.10.11; platform_python_implementation != 'PyPy'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.0.0; python_version >= '3.9'
portalocker==2.10.1; python_version >= '3.8'
propcache==0.2.0; python_version >= '3.8'
protobuf==5.28.3; python_version >= '3.8'
pyclipper==1.3.0.post6
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.1; python_full_version < '3.12.4'
pydantic-core==2.27.1; python_version >= '3.8'
pyjwt[crypto]==2.10.0; python_version >= '3.9'
pymupdf==1.24.14; python_version >= '3.9'
pymysql==1.1.1; python_version >= '3.7'
pypdf==5.1.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
rapidocr-onnxruntime==1.4.0; python_version < '3.13' and python_version >= '3.6'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
shapely==2.0.6; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.36; python_version >= '3.7'
sympy==1.13.3; python_version >= '3.8'
tenacity==8.5.0; python_version >= '3.8'
tiktoken==0.8.0; python_version >= '3.9'
tqdm==4.67.0; python_version >= '3.7'
types-requests==2.32.0.20241016; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2024.2; python_version >= '2'
urllib3==2.2.3; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'
yarl==1.18.0; python_version >= '3.9'
